pyproject.toml
[tool.poetry]
name = "calcio_splash"
version = "0.1.0"
description = ""
authors = ["Stella Margonar <margonar@spaziodati.eu>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0.4"
django-storages = "^1.12.3"
django-dbbackup = "^3.3.0"
boto3 = "^1.23.4"
raven = "^6.10.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
Fabric3 = "^1.14.post1"
pre-commit = "^2.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 119
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
skip_glob = ["*/node_modules/*.py", "*/migrations/*.py",]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "DJANGO_THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_django = ["django"]

[tool.black]
line-length = 119
skip-string-normalization = true
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.venv
  | migrations
  | node_modules
)/
'''